FEAT: Easy dotfile management with .stowrc file
Basically, what was previously done, was that each directory was stowED
separately, and the directory structure was not well organised as there
was a .config directory in ever single directory (except zsh obv).

What I did now is added a .stowrc file and told it the target directory.
Now if we run the stow . command while being in the dotfiles directory,
it would automatically stow every single directory to ~/.config
directory. (I KNOW I AM SAYING DIRECTORY TOO MANY TIMES)

It would not do such with the zsh directory as it is ignored because the
.zshrc and .zshenv files have to live in ~ and not in ~/.config. Thus,
there is a separate command to stow it, in the README.md file.

This is a wayyy better use of stow. Just with the power of one file
(.stowrc), the dotfile management get better.

I was also doing the same thing you are doing and when I got to know
this, I applied the change and it was magnificent. So I decided to help
you too. PWEASE GIMME A SHOUTOUT ON YOUR YT CHANNEL. I am 15 years old
:)